backend/requirements.txt
Flask
Flask-CORS
web3
python-dotenv
redis
